requirements.txt
django==5.2.1
django-debug-toolbar==5.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pycountry==24.6.1
requests==2.32.3
tenacity==9.1.2
schedule==1.2.2
uvicorn[standard]==0.34.2
beautifulsoup4==4.13.0
debugpy==1.8.14